rename Account interface to OwnedAccount (#55)
The Account type flip found
[here](https://github.com/onflow/flips/pull/92) is set to define a new
type which will cause a conflict in our own type definitions.
AuthAccount is going to be split into a set of interfaces and a concrete
called `Account` which we also define. If the flip were to be
implemented, that type would brick our own contract due to `Account`
being a reserved type/keyword.

Since we aren't in mainnet yet, we can rename the type here and avoid
the conflict entirely.