Update CapabilityFactory Manager (#84)
Closes: #73 
Coverage: 75.6%

Added pre condition to `Manager.addFactory()` to prevent overwriting
existing factories. Also added `Manager.updateFactory()` which
overwrites Factories on given Type, and adds if it didn't already exist.
The thought here is that the behavior is more explicit and provides a
safer method to add new Factory types depending on developer preference,
but open to feedback on this behavior.

Also added comments to CapabilityFactory contract.

Last note, I added a delimiter `'_'` to derived canonical paths between
the prefix and deployment Address. I thought this would make it easier
to split the strings should the need arise.