Fix owner assignment (#86)
Closes: #65
Coverage: 75.2%

Assigning owner similar to pattern for assigning `ProxyAccount` parent
with `setOwnerCallback(_ addr: Address)` added to `OwnedAccount`
interface. This callback both sets the `ChildAccount.acctOwner` as the
address of the calling `Manager` and clears the new
`ChildAccount.pendingOwner` field. This new field enables onchain state
to reflect the in-flight transfer of ownership status and ensure that
the correct user is set as the owner.